adding fixes to fetch and moving stuff to files
